requirements.txt
aiohttp==3.8.4
aiomysql==0.1.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.0.1
chat-exporter==2.5.3
discord.py==2.1.1
emoji==2.2.0
frozenlist==1.3.3
grapheme==0.6.0
idna==3.4
multidict==6.0.4
NBT==1.5.1
numerize==0.12
PyMySQL==1.0.2
python-dotenv==1.0.0
pytz==2022.7.1
yarl==1.8.2
matplotlib==3.7.0
Pillow~=9.5.0
